% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getYearly.R
\name{getYearly}
\alias{getYearly}
\title{指定した年の株式情報を取得}
\usage{
getYearly(CODE, YEAR)
}
\arguments{
\item{CODE}{numeric}

\item{YEAR}{numeric}
}
\value{
data.frame 過去 300 日分の株式情報
}
\description{
指定した年の株式情報を取得
}
